+++ make/photon/db/registry_sqlite.sql
@@ -261,6 +261,26 @@ create table harbor_label (
  UNIQUE(name, scope)
  );
 
+create table harbor_resource_label (
+ id INTEGER PRIMARY KEY,
+ label_id int NOT NULL,
+/*
+ the resource_id is the ID of project when the resource_type is p
+ the resource_id is the ID of repository when the resource_type is r
+ the resource_id is the name of image when the resource_type is i
+*/
+ resource_id varchar(256) NOT NULL,
+/*
+ 'p' for project
+ 'r' for repository
+ 'i' for image
+*/
+ resource_type char(1) NOT NULL,
+ creation_time timestamp default CURRENT_TIMESTAMP,
+ update_time timestamp default CURRENT_TIMESTAMP,
+ UNIQUE (label_id,resource_id, resource_type)
+ );
+
 create table alembic_version (
     version_num varchar(32) NOT NULL
 );